// questionaire_frontend/src/utils/models/Customer.ts
export interface Customer {
    id: number;
    name: string;
    contact_name: string;
    email: string;
    phone: string;
    job_title: string;
    city: string;
    street: string;
    postalcode: string;
    state: string;
    county: string;
    country: string;
    latitude: string;
    longitude: string;
    fax: string;
    business_type: string;
    created_at: string;
    updated_at: string;
}